INET: fix crash recovery script
All,

This fixes the recovery script for INET so it actually recovers. At the
bottom I've added an example (uhu!) of how to add 3rd party daemons to
this script that need to be recovered upon restart of INET. I think we
all run SSHd, so it's sort of convenient to have.

--
Thomas

>From ad59b319afa18c3bb38efae4b76692ba3f8c69d4 Mon Sep 17 00:00:00 2001
From: Thomas Veerman <thomas@minix3.org>
Date: Thu, 13 Sep 2012 11:51:36 +0000
Subject: [PATCH 4/4] INET: fix crash recovery script