kernel: handle pagefaults in vm_memset()
All,

I really like the look of Arne's /dev/zero changes; the commit separation
and coding style is textbook. A new kernel call is somewhat unnecessarily
heavyweight but as he's done the work and it looks clean I think it's worth
accepting.

Somewhat to my surprise, and entirely clinching the deal for me, I
measure a 0.9% performance improvement for 'make servers'.. There are some
zeroes being produced for them apparently ;)