arm:Add spin to syslib on arm.
NOTE:read_tsc is still not functional at this point. so the code
will never exit on timeout.

libsys:don't use read_tsc in spin functions.

We are lacking a framework for user accessible high resolution
timers on arm and will therefore use system calls to get the
current uptime.

Change-Id: Ifed387219904d2a9781c8c5a609cf38733d8eb32