vm: merge i386 and arm pagetable code
All,

This patch merges the arm and i386 pagetable code, at least into a
single file.. arch-specific header files are merged into one, and
some previously-arch-specific stuff is arch-independent. The remaining
differences are a bunch of #if-tested statements in pagetable.c.

The resulting file is quite noisy but this way I can keep reducing
the differences and not risking them diverging. Assuming I didn't
break the ARM version (big assumption :) - I'll check with the ARM
boys) it's a net improvement to merge it imho.

Change-Id: I98da4a4dc17a03db4e7209366c742fc693163abc