mkfs, mkproto: minor improvements
The first in a series of simplifications / improvements to mkfs.

The only real end goal is to implement indirect blocks in directories
so that we can make 4k-block-sized filesystems but I fixed/added a few
more things along the way.

A full Minix disk image generated with mkproto / mkfs.mfs is *much* smaller
now, with 4k blocks 64MB uncompressed when built dynamically linked.

	. mkfs: -x feature for extra space
	. rename manpage to mkfs.mfs.1 to follow the binary
	. move mkproto so it can become part of the tools easily