VFS: fix deadlock when out of worker threads
All,

There is a deadlock vulnerability when there are no worker threads
available and all of them blocked on a worker thread that's waiting for
a reply from a driver or a reply from an FS that needs to make a back
call. In these cases the deadlock resolver thread should kick in, but
didn't in all cases. Moreover, POSIX calls from File Servers weren't
handled properly anymore, which also could lead to deadlocks.

--
Thomas

commit badec36b331e6422762876434086f6c897a38188
Author: Thomas Veerman <thomas@minix3.org>
Date:   Wed Nov 14 13:12:37 2012 +0000

    VFS: fix deadlock when out of worker threads

    There is a deadlock vulnerability when there are no worker threads
    available and all of them blocked on a worker thread that's waiting for a
    reply from a driver or a reply from an FS that needs to make a back call. In
    these cases the deadlock resolver thread should kick in, but didn't in all
    cases. Moreover, POSIX calls from File Servers weren't handled properly
    anymore, which also could lead to deadlocks.