vm: replace phys avl by array
Dear all,

A change inspired by profiling other changes I'm working on; VM was
doing a lot of avl tree manipulation for physical blocks.  But now
that physical blocks are always one page, we can use an array
instead, which takes about the same amount of memory as long as the
memory isn't too sparse, hence the default stack size reduction. And
it's always nice to gets rid of some code.

VM uses the libminc malloc() for it, so it has to implement its own
mmap() and brk() to get memory.

This can also replace the custom slab allocator, to throw out some
more code, but as there is no measurable performance difference I
didn't think it was worth the change.

This makes 'make' in servers/ about 3% faster.

 . make vm be able to use malloc() by overriding brk()
   and minix_mmap() functions
 . phys regions can then be malloc()ed and free()d instead
   of being in an avl tree, which is slightly faster
 . 'offset' field in phys_region can go too (offset is implied
   by position in array) but leads to bigger code changes