libc, libutil: remove compat hacks
I spent a lot of time debugging why the dynamically linked unzip
crashes. It turns out it's because of the __RENAME()d symbol hack -
the jumps are wrong for the PIC case after all. They get exercised
when an application calls a __RENAME()d symbol without including
the prototype, i.e. the old version in netbsd or the compat hacked
version in minix will be called.

I weighed a few options (make the call PIC-compatible, make __RENAME()
a noop for minix, add aliases manually) but finally decided just
making them all in-source aliases for once and for all was the
easiest, safest, cleanest way out.. (And we have to keep the
__RENAME()d symbols for ssh/sshd which is a dynamically linked
binary in the wild already..!) I hope you agree. It gets rid of the
messy hack at the cost of a few small minix-specific lines in libc,
and the (imho minor) maintenance cost of having to add or edit a
few in the future.