Fields and functions for new 2ndary cache.
This commit introduces message fields, library calls and their calls
in libminxfs for using the new secondary cache, without actually
including that functionality yet. This is to make bisecting problems
that might not be due to the new cache code and reading the patches
easier, and forces the fallback in libminixfs to be working in the
absence of the secondary cache functionality.

Change-Id: I76fc5242864db63d134f86d1ce1446b0072ea337