Bump version to 3.3.0
We're bumping the version to 3.3.0. This needed to be done anyway,
but it is also as a preparation for the next few commits that will
make struct stat identical to NetBSD's. In general all changes will
be fully backwards compatible, except for dynamically linked
binaries. By changing the version number we can provide new versions
of those binaries that are built for the new struct stat/off_t/ino_t.

Change-Id: Idc200bb6f75cdce7604e7706a5e30518496e543b