arm:mmc fixes
When we send MMC commads that contain data the controller provide no
description of the order of events and we need to be able to send data
from and to the controller "when needed". Changed the code to react
on buffer read and buffer write ready based on interrupts.

We also increase driver process priority.  Without special attribute
drivers will have the same priority as other userland processes this.
Posix test 1 was taking so much resources that it triggered alarms
in the MMC driver.

Change-Id: I60c9140bf0e45b74be6475054564d4e1bd89f21e