testfs: test MFS format doesn't change, using mkfs
This test tests mkfs.mfs will generate the same FS image given the same
input files. mkproto creates a proto file (normalizing directory entry
order). The assumption is that a change in the output flags a tacit
change in FS format, and that a FS format change will cause the image
to change.

	. Changes to mkfs.mfs that innocently change the format can
	  change the sha1 output in the script along with it.
	. The assumption is that corresponding versions of mkfs.mfs and
	  MFS will always work together; otherwise a lot breaks (ramdisk etc.)
	. Therefore, as long as a generated FS image stays the same with the
	  same input now, incompatible MFS changes will still be flagged,
	  even if they work together with the current mkfs.mfs.

Change-Id: I120cd1d9347aedaf438c8847f0e7b8eccf086e9d