kernel:WIP pending changes dynamic configuration.
Change-Id: I0f0ef014e5f341ad5df8c048863c8e5032b80950

arm:switch kernel to the uImage format

uImage format is the native image format used in u-boot. Switching
allows us to steamline the boot process a little. It also  allows
us to no longer need to read the command line arguments as a separate
module as we can now set the u-boot bootargs variable.

remove references to cmdline.txt

Kernel WIP export "board" as the minix board name.

Fixes for the kernel