ALIX does not have a device at bus 0, device 0, function 0
Seems like its a kind of convention introduced by Intel but not mandatory from a PCI specification point of view, that the PCI host controller resides on bus 0, device 0 function 0. At least tinybios (the bios used by ALIX and WRAP) based appliances are not able to boot with this change.