x86 iso: overhaul creation of installation CD
The CD now boots directly from the ISO 9660 filesystem instead of using
MBR partitioning with Minix file systems. This saves some space on the
CD and reduces memory requirements by some unknown amount as the root
ramdisk is completely eliminated.

The setup is also modified to be more NetBSD-like (unpacking sets
tarballs instead of blindly copying the CD contents). This allows
massive space savings because only the base system used for installation
is stored uncompressed and more modularity (to be exploited at a later
time). The sets themselves are to be refined, but are enough for now.

Those two steps are melded together because copying the CD contents with
isofs as the root filesystem results in some failed tests in
minix-posix, probably due to some obscure corner-case bugs in isofs.

Having a read-only root filesystem outputs some error messages because
it isn't properly supported. Ramdisks on /var and /tmp are used to
mitigate this, but the clean solution would involve union mounts,
something Minix 3 doesn't support for the time being.

Change-Id: Ib4700d9f10c5b4b297e0994b5338847951fc79b4